frecon: Fix type casting in image_show
BUG=b:205262466
TEST=Verify that powerwash screen is shown

Known Issue: Screen will be cropped

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/frecon/+/3309666
Tested-by: Dominik Behr <dbehr@chromium.org>
Tested-by: Bob Moragues <moragues@chromium.org>
Commit-Queue: Bob Moragues <moragues@chromium.org>
(cherry picked from commit 8e57ac3d75cc4b0351fae3067693527fbcf68cee)

+++ image.c
@@ -183,8 +183,8 @@ int image_show(image_t* image, fb_t* fb)
 		startx = image->location_x;
 		starty = image->location_y;
 	} else {
-		startx = (fb_getwidth(fb) - w)/2;
-		starty = (fb_getheight(fb) - h)/2;
+		startx = (fb_getwidth(fb) - (int32_t)w)/2;
+		starty = (fb_getheight(fb) - (int32_t)h)/2;
 	}
 
 	if (image->use_offset) {
